SCT_Monitoring: Fix clang warnings.
Clang warnings: unused private data members; braces in initializers;
                unused parameters; int/float conversion issues.



Former-commit-id: 9544844bb7d4a547a8a1c95f5b34734395aedfad

+++ InnerDetector/InDetMonitoring/SCT_Monitoring/SCT_Monitoring/SCTHitsNoiseMonTool.h
@@ -498,9 +498,9 @@ class SCTHitsNoiseMonTool : public SCTMotherTrigMonTool{
   Prof2_t
     prof2Factory(const std::string & name, const std::string & title, const SCT_Monitoring::Bec bec, MonGroup & registry, VecProf2_t & storageVector);
   Prof2_t
-    prof2DFactory(const std::string & name, const std::string & title, MonGroup & registry, int nbinx, int xlo, int xhi, int nbiny, int ylo, int yhi);
+    prof2DFactory(const std::string & name, const std::string & title, MonGroup & registry, int nbinx, float xlo, float xhi, int nbiny, float ylo, float yhi);
   Prof_t
-    profFactory(const std::string & name, const std::string & title, MonGroup & registry, int nbin, int lo, int hi);
+    profFactory(const std::string & name, const std::string & title, MonGroup & registry, int nbin, float lo, float hi);
   Prof_t
     profFactory(const std::string & name, const std::string & title, MonGroup & registry);
   //@}
